Add support for custom es routing
Add a CustomRouting interface that exposes a getRoutingKey method to be implemented by developers.
The routing key provided by this method is then provided to Documents and used in index and delete operations.

This pr is linked to the elastic-adapter pr https://github.com/babenkoivan/elastic-adapter/pull/9 and we should first merge the latter.
The changes on elastic-adapter should be backwards compatible and so no major version upgrade should be required.